Avoid extra entries in conandata.
When export with conan export it will include all data in it, including
that of all other package versions. That dirties all packages when a
new version is added. This change resolves that issue by injecting
a hook at export time that cleans the conandata to only include the
relevant version fields.

fixes https://github.com/bfgroup/barbarian/issues/16